requirements.txt
alembic
argon2-cffi
argon2-cffi-bindings
boto3
botocore
certifi
cffi
colorama
graphviz
greenlet
halo
jmespath
log-symbols
Mako
MarkupSafe
minio
numpy
pandas
parquet_tools
psycopg2-binary
pyarrow
pycparser
pycryptodome
python-dateutil
python-dotenv
pytz
s3transfer
six
spinners
SQLAlchemy
tabulate
termcolor
thrift
typing_extensions
tzdata
urllib3
